+net:ethernet.h - выделил в него описание езернет MACадресов. *   :mac - оптимизировал для работы в МИПС невыравненые адреса mac_addr_float работают с магией компилятора -                 для работы с packed gcc использует lwl/lwr, swl/swr инструкции
